removed warnings and improved tests
