--- src-tauri/utils/file_utils/src/CryptoFiles.rs
@@ -1,2 +1,0 @@
-pub mod CryptoFiles;
-pub mod Parser;
