+++ src-tauri/utils/file_utils/src/crypto_files.rs
@@ -1,0 +1,3 @@
+pub mod crypto_files;
+pub mod parser;
+pub mod file_traversal;
